+++ client/main.scss
@@ -41,5 +41,7 @@
 @import '{}/imports/ui/components/report/report';
 @import '{}/imports/ui/components/sharing/sharing';
 
+@import '{}/imports/ui/pages/calendar/calendar';
+
 // Stylesheets from old structure outside of imports
 @import 'views/views';
