+++ client/views/_views.scss
@@ -1,9 +1,6 @@
 // Views stylesheet imports
 // --------------------------------------------------
 
-// Calendar
-@import 'calendar/calendar';
-
 // Courses
 @import 'courses/compact/course.compact';
 @import 'courses/details/course.details';
