+++ server/extracts.msgfmt~
@@ -1,1 +1,1 @@
-[{"event.list.windowtitle":{"key":"event.list.windowtitle","text":"Events","file":"client/views/kiosk/kiosk.js","line":55,"mtime":1509461520066,"template":"unknown","ctime":1509461520066},"languageSelection.noLanguagesFound":{"key":"languageSelection.noLanguagesFound","text":"No languages found.","file":"imports/ui/components/languages/languages.html","line":33,"mtime":1502143991304,"template":"languageSelectionWrap"}},{"extractedAt":1509461656942,"updatedAt":1509461520066}]
+[{"calendar.next.month.long":{"key":"calendar.next.month.long","text":"Go to next month","file":"imports/ui/pages/calendar/calendar.html","line":107,"mtime":1509028298856,"template":"calendar"},"calendar.next.month.short":{"key":"calendar.next.month.short","text":"next month","file":"imports/ui/pages/calendar/calendar.html","line":101,"mtime":1509028298856,"template":"calendar"},"calendar.next.week.long":{"key":"calendar.next.week.long","text":"Go to next week","file":"imports/ui/pages/calendar/calendar.html","line":106,"mtime":1509028298856,"template":"calendar"},"calendar.next.week.short":{"key":"calendar.next.week.short","text":"next week","file":"imports/ui/pages/calendar/calendar.html","line":100,"mtime":1509028298856,"template":"calendar"},"calendar.next.year.long":{"key":"calendar.next.year.long","text":"Go to next year","file":"imports/ui/pages/calendar/calendar.html","line":108,"mtime":1509028298856,"template":"calendar"},"calendar.next.year.short":{"key":"calendar.next.year.short","text":"next year","file":"imports/ui/pages/calendar/calendar.html","line":102,"mtime":1509028298856,"template":"calendar"},"calendar.noCoursesThisDay":{"key":"calendar.noCoursesThisDay","text":"There are no events on this day","file":"imports/ui/pages/calendar/calendar.html","line":39,"mtime":1509028298856,"template":"calendar"},"calendar.previous.month.long":{"key":"calendar.previous.month.long","text":"Go to previous month","file":"imports/ui/pages/calendar/calendar.html","line":104,"mtime":1509028298856,"template":"calendar"},"calendar.previous.month.short":{"key":"calendar.previous.month.short","text":"previous month","file":"imports/ui/pages/calendar/calendar.html","line":98,"mtime":1509028298856,"template":"calendar"},"calendar.previous.week.long":{"key":"calendar.previous.week.long","text":"Go to previous week","file":"imports/ui/pages/calendar/calendar.html","line":103,"mtime":1509028298856,"template":"calendar"},"calendar.previous.week.short":{"key":"calendar.previous.week.short","text":"previous week","file":"imports/ui/pages/calendar/calendar.html","line":97,"mtime":1509028298856,"template":"calendar"},"calendar.previous.year.long":{"key":"calendar.previous.year.long","text":"Go to previous year","file":"imports/ui/pages/calendar/calendar.html","line":105,"mtime":1509028298856,"template":"calendar"},"calendar.previous.year.short":{"key":"calendar.previous.year.short","text":"previous year","file":"imports/ui/pages/calendar/calendar.html","line":99,"mtime":1509028298856,"template":"calendar"},"calendar.weekNumber":{"key":"calendar.weekNumber","text":"Week ","file":"imports/ui/pages/calendar/calendar.html","line":57,"mtime":1509028298856,"template":"calendar"},"calendar.windowtitle":{"key":"calendar.windowtitle","text":"Calendar","file":"imports/startup/client/routes.js","line":38,"mtime":1509462426013,"template":"unknown"},"course.windowtitle":{"key":"course.windowtitle","text":"Course: {COURSE}","file":"imports/startup/client/routes.js","line":172,"mtime":1509462426013,"template":"unknown","ctime":1509462426013},"profile.settings.windowtitle":{"key":"profile.settings.windowtitle","text":"My Profile Settings - {USER}","file":"imports/startup/client/routes.js","line":232,"mtime":1509462426013,"template":"unknown","ctime":1509462426013},"profile.windowtitle":{"key":"profile.windowtitle","text":"{USER}\\","file":"imports/startup/client/routes.js","line":76,"mtime":1509462426013,"template":"unknown","ctime":1509462426013},"venue.edit.siteTitle.create":{"key":"venue.edit.siteTitle.create","text":"Create Venue","file":"imports/startup/client/routes.js","line":316,"mtime":1509462426013,"template":"unknown","ctime":1509462426013}},{"extractedAt":1509462504343,"updatedAt":1509462426013}]
